refactor: convert engines.py into a package with focused submodules
Moves shared computation logic from loan/engines.py into the new
money_warp/engines/ package (interest.py, allocation.py, fines.py,
forward_pass.py). Both Loan and BillingCycleLoan now import shared
engine functions from money_warp.engines instead of reaching across
product boundaries.

loan/engines.py is kept as a backward-compatible re-export shim.
loan/__init__.py uses PEP 562 lazy __getattr__ for Loan to break
the circular import chain.

Closes #44
