+++ money_warp/loan/__init__.py
@@ -1,17 +1,21 @@
 """Loan module for personal loan modeling with flexible payment schedules."""
 
 from .allocation import Allocation
-from .engines import InterestCalculator, MoraStrategy
 from .installment import Installment
-from .loan import Loan
 from .settlement import AnticipationResult, Settlement
 
 __all__ = [
     "Allocation",
     "AnticipationResult",
     "Installment",
-    "InterestCalculator",
     "Loan",
-    "MoraStrategy",
     "Settlement",
 ]
+
+
+def __getattr__(name: str):  # type: ignore[misc]
+    if name == "Loan":
+        from .loan import Loan
+
+        return Loan
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
