+++ money_warp/loan/loan.py
@@ -5,14 +5,7 @@ from datetime import date, datetime
 from typing import Dict, List, Optional, Type
 
 from ..cash_flow import CashFlow, CashFlowItem, CashFlowType
-from ..interest_rate import InterestRate
-from ..money import Money
-from ..rate import Rate
-from ..scheduler import BaseScheduler, PaymentSchedule, PaymentScheduleEntry, PriceScheduler
-from ..tax.base import BaseTax, TaxResult
-from ..time_context import TimeContext
-from ..tz import to_datetime, tz_aware
-from .engines import (
+from ..engines import (
     InterestCalculator,
     LoanState,
     MoraStrategy,
@@ -22,6 +15,13 @@ from .engines import (
     covered_due_date_count,
     is_payment_late,
 )
+from ..interest_rate import InterestRate
+from ..money import Money
+from ..rate import Rate
+from ..scheduler import BaseScheduler, PaymentSchedule, PaymentScheduleEntry, PriceScheduler
+from ..tax.base import BaseTax, TaxResult
+from ..time_context import TimeContext
+from ..tz import to_datetime, tz_aware
 from .installment import Installment
 from .settlement import AnticipationResult, Settlement
 from .tvm import loan_calculate_anticipation, loan_irr, loan_present_value
